Refactor enrollment's ExportCSV
Adds a new type to contain all the export information so the call is
still straightforward, but the caller still gets to decide if the export
CSV and database record are kept.  Allows removing these pieces for the
Canvas call in case the DB/filesystem operations succeed, but the HTTP
request ends up failing.